requirements.txt
keras==2.11.0
matplotlib==3.6.2
numpy==1.24.1
Pillow==9.4.0
Scrapy==2.7.1
tensorflow==2.11.0
tensorflow_intel==2.11.0
